LayoutTests/imported/w3c/web-platform-tests/content-security-policy/unsafe-hashes/style_attribute_denied_wrong_hash-expected.txt


FAIL Test that the inline style attribute is blocked assert_equals: expected "style-src-attr" but got "style-src"
